+++ app/ts/controllers/NegociacaoController.ts
@@ -19,7 +19,7 @@ class NegociacaoController {
         const negociacao = new Negociacao(
             new Date(this._inputData.value.replace(/-/g, ',')),
             parseInt(this._inputQuantidade.value),
-            parseInt(this._inputValor.value)
+            parseFloat(this._inputValor.value)
         );
 
         console.log(negociacao);
